refactor: re-organize types
