+++ frontend/src/types/Studio/StudioResource.ts
